services/prompt_builder/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
uvloop>=0.16.0
httptools>=0.4.0
python-multipart>=0.0.5,<0.1.0
requests>=2.26.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=0.19.0,<0.20.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
tiktoken>=0.5.1 